#91 add css rules for search results component relating to filter bar
This PR adds a new CSS file and the required clientlibs wiring required to provide some very basic / lightweight styling to new search results related html elements.

I made an attempt to stick with the basic structure / design of existing elements and the new CSS rules attempt to leave the basic look and feel of all elements on the page untouched.

Note that this PR only makes sense in the context of another PR being made against the `aem-core-cif-components` repo, though the changes should be reasonably safe independently.